Move vertical tab related tests into its directory
