+++ browser/ui/views/frame/brave_tabs_search_button_browsertest.cc
@@ -5,9 +5,6 @@
 
 #include "brave/browser/ui/brave_browser.h"
 #include "brave/browser/ui/browser_commands.h"
-#include "brave/browser/ui/views/frame/brave_browser_view.h"
-#include "brave/browser/ui/views/frame/vertical_tabs/vertical_tab_strip_region_view.h"
-#include "brave/browser/ui/views/frame/vertical_tabs/vertical_tab_strip_widget_delegate_view.h"
 #include "brave/browser/ui/views/tabs/vertical_tab_utils.h"
 #include "brave/components/constants/pref_names.h"
 #include "build/build_config.h"
@@ -24,7 +21,6 @@
 #include "chrome/test/base/ui_test_utils.h"
 #include "components/prefs/pref_service.h"
 #include "content/public/test/browser_test.h"
-#include "ui/views/test/button_test_api.h"
 
 class BraveTabsSearchButtonTest : public InProcessBrowserTest,
                                   public ::testing::WithParamInterface<bool> {
@@ -62,54 +58,3 @@ INSTANTIATE_TEST_SUITE_P(BraveTabsSearchButtonTest,
                          BraveTabsSearchButtonTest,
                          ::testing::Bool());
 #endif
-
-class VerticalTabSearchButtonBrowserTest : public InProcessBrowserTest {
- public:
-  BraveBrowserView* browser_view() {
-    return BraveBrowserView::From(
-        BrowserView::GetBrowserViewForBrowser(browser()));
-  }
-
-  TabSearchButton* tab_search_button() {
-    return browser_view()
-        ->vertical_tab_strip_widget_delegate_view()
-        ->vertical_tab_strip_region_view()
-        ->GetTabSearchButtonForTesting();
-  }
-
-  TabSearchBubbleHost* tab_search_bubble_host() {
-    return browser_view()->GetTabSearchBubbleHost();
-  }
-
-  WebUIBubbleManager* bubble_manager() {
-    return tab_search_bubble_host()->webui_bubble_manager_for_testing();
-  }
-
-  void RunUntilBubbleWidgetDestroyed() {
-    ASSERT_NE(nullptr, bubble_manager()->GetBubbleWidget());
-    base::RunLoop run_loop;
-    base::SingleThreadTaskRunner::GetCurrentDefault()->PostTask(
-        FROM_HERE, run_loop.QuitClosure());
-    run_loop.Run();
-    ASSERT_EQ(nullptr, bubble_manager()->GetBubbleWidget());
-  }
-};
-
-IN_PROC_BROWSER_TEST_F(VerticalTabSearchButtonBrowserTest,
-                       ButtonClickCreatesBubble) {
-  brave::ToggleVerticalTabStrip(browser());
-  ASSERT_TRUE(tabs::utils::ShouldShowVerticalTabs(browser()));
-
-  ASSERT_EQ(nullptr, bubble_manager()->GetBubbleWidget());
-
-  ui::MouseEvent dummy_event(ui::MouseEvent(ui::EventType::kMousePressed,
-                                            gfx::PointF(), gfx::PointF(),
-                                            base::TimeTicks::Now(), 0, 0));
-  views::test::ButtonTestApi(tab_search_button()).NotifyClick(dummy_event);
-  ASSERT_NE(nullptr, bubble_manager()->GetBubbleWidget());
-
-  tab_search_bubble_host()->CloseTabSearchBubble();
-  ASSERT_TRUE(bubble_manager()->GetBubbleWidget()->IsClosed());
-
-  RunUntilBubbleWidgetDestroyed();
-}
